Move analysis to a more logical location
