+++ unittests/IR/ForbiddenPreconditionInfluence.cpp
@@ -2,17 +2,16 @@
 //
 // Part of the LLZK Project, under the Apache License v2.0.
 // See LICENSE.txt for license information.
-// Copyright 2025 Veridise Inc.
+// Copyright 2026 Project LLZK
 // SPDX-License-Identifier: Apache-2.0
 //
 //===----------------------------------------------------------------------===//
 
-#include "llzk/Analysis/ForbiddenPreconditionInfluence.h"
+#include "llzk/Dialect/Verif/Util/ForbiddenPreconditionInfluence.h"
 
 #include "../LLZKTestBase.h"
 #include "../LLZKTestUtils.h"
 
-#include "llzk/Analysis/AnalysisUtil.h"
 #include "llzk/Dialect/Bool/IR/Ops.h"
 #include "llzk/Dialect/Felt/IR/Ops.h"
 #include "llzk/Dialect/Function/IR/Ops.h"
@@ -31,6 +30,7 @@
 using namespace mlir;
 using namespace llzk;
 using namespace llzk::component;
+using namespace llzk::verif;
 
 class ForbiddenPreconditionInfluenceTests : public LLZKTest {};
